apps/feature-crawler/requirements.txt
playwright==1.48.0
kafka-python==2.0.2
tldextract==5.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
Pillow==10.4.0
imagehash==4.3.1
pytesseract==0.3.10
regex==2024.9.11
idna==3.7
ujson==5.10.0
PyYAML==6.0.2
requests==2.32.3